refactor(desktop): DRY cloud sync — extract CloudProviderCard + useCloudProviders
CloudSyncSetupDialog and MobileSyncTab were each maintaining their own copies
of the provider SVG icons, label/detail strings, card layout, and
connect/disconnect logic. This caused the GDrive/Dropbox logos to go missing
from Settings when the two implementations drifted apart.

Extract two shared primitives:

- hooks/useCloudProviders.ts: single state machine for connect/disconnect.
  Handles initiateDesktopOAuth → storeCloudToken → startCloudSync, error
  state, and clearCloudProvider. Both consumers import this; neither
  re-implements the flow inline.

- components/CloudProviderCard.tsx: single source of truth for the provider
  card UI — icons (Dropbox first, then GDrive), name, detail copy, and
  connect/disconnect buttons. The onDisconnect prop is optional so the
  onboarding dialog (connect-only) and the Settings tab (connect + disconnect)
  can both use the same component.

Both CloudSyncSetupDialog and MobileSyncTab are now thin wrappers.

Add components/COMPONENTS.md: a component index that lists every desktop
component and its purpose. AI and humans must check this before creating new
components to prevent future duplication.
